// src/ct-types.d.ts
export type Address = AddressUpdate & {
  /**
   * @deprecated
   */
  geoLat: string | null
  /**
   * @deprecated
   */
  geoLng: string | null
  latitudeLoose: string | null
  longitudeLoose: string | null
  /**
   * @deprecated
   */
  markerColor: string | null
  /**
   * @deprecated
   */
  markerIcon: string | null
  /**
   * @deprecated
   */
  markerUrl: string | null
  /**
   * @deprecated
   */
  meetingAt: string | null
  /**
   * @deprecated
   */
  postalcode: string | null
}
// ...
export type AddressUpdate = AddressCreate & {
  id: number
}
// ...
export type AppointmentAdditionals = {
  date: DateString
  id: number
  isRepeated: boolean
  meta: MetaModified
}

export type AppointmentBase = {
  additionals: Array<AppointmentAdditionals>
  /**
   * Use 'additionals' instead
   * @deprecated
   */
  additions?: Array<AppointmentAdditionals>
  address: Address
  allDay: boolean
  calendar: Calendar
  /**
   * Use 'title' instead
   * @deprecated
   */
  caption?: string
  description: string | null
  endDate: ZuluDate | DateString
  exceptions: Array<AppointmentExeptions>
  id: number
  image: Image | null
  /**
   * Use 'description' instead
   * @deprecated
   */
  information?: string | null
  isInternal: boolean
  link: string | null
  meta: MetaDataEntityId
  /**
   * Use 'subtitle' instead
   * @deprecated
   */
  note?: string | null
  onBehalfOfPid: number | null
  repeatFrequency: number | null
  /**
   * ID of the repeat pattern, NONE = 0, DAILY = 1, WEEKLY = 7, MONTHLY_BY_DATE = 31, MONTHLY_BY_WEEKDAY = 32, YEARLY = 365, MANUALLY = 999
   */
  repeatId: 0 | 1 | 7 | 31 | 32 | 365 | 999
  repeatOption: number | null
  repeatUntil: DateStringNullable
  signup: AppointmentSignup | null
  startDate: ZuluDate | DateString
  subtitle: string | null
  title: string
  version: number
}

export type AppointmentBooking = {
  id?: number
  minpost?: string
  minpre?: string
  resource_id?: string
  status_id?: number
}

export type AppointmentCalculated = {
  base: AppointmentBase
  calculated: {
    endDate: ZuluDate
    startDate: ZuluDate
  }
}
// ...
export type AppointmentExeptions = {
  date: string
  id: number
  meta: MetaDataEntityId
}

export type AppointmentSignup = {
  signupDaysArchiveGroupNo: number | null
  signupGroupTypeId: number | null
  signupSetCompletionDate: boolean
  signupTemplateGroupId: number | null
  singupDaysForwardNo: number | null
}
// ...
export type Calendar = CalendarCreate & {
  id: number
  /**
   * @deprecated
   */
  isPrivate?: boolean
  /**
   * @deprecated
   */
  isPublic?: boolean
  meta: MetaModified
  nameTranslated: string
  randomUrl: string
}

export type CalendarCreate = CalendarUpdate & {
  type: 'church' | 'group' | 'personal'
}
// ...
export type DateString = string

/**
 * A simple date in ISO format, e.g. '2022-10-19'
 */
export type DateStringNullable = string | null
// ...
export type Image = {
  additionalInfos: Array<string>
  domainId: string
  domainType: string
  fileUrl: string
  filename: string
  id: number
  imageOption: {
    crop?: {
      bottom?: string
      left?: string
      right?: string
      top?: string
    }
    focus?: {
      x?: string
      y?: string
    }
  }
  imageUrl: string
  meta: MetaDataEntityId
  name: string
  relativeUrl: string
  securityLevelId: number | null
  showOnlyWhenEditable: boolean
  size: number | null
  type: string
}
// ...
export type MetaDataEntityId = {
  createdDate: ZuluDate
  createdPerson: {
    id: number
  }
  modifiedDate?: ZuluDate
  modifiedPerson?: {
    id: number
  }
}
// ...
export type MetaModified = {
  modifiedDate?: ZuluDate
  modifiedPid?: number
}
// ...
export type ZuluDate = `${number}-${number}-${number}T${number}:${number}:${number}Z`